requirements.txt
streamlit
requests
beautifulsoup4
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-groq
tiktoken
python-dotenv
